ct-cinematic-viewer/requirements.txt
numpy>=1.21.0
pydicom>=2.3.0
SimpleITK>=2.2.0
scikit-image>=0.19.0
trimesh>=3.12.0
torch>=1.12.0
tqdm>=4.62.0
pillow>=9.0.0 